Rename RegistryService to SteamRegistry, and update namespace
