+++ DoomLauncher/DoomLauncher.csproj
@@ -286,7 +286,7 @@
     <Compile Include="Handlers\Files\IFileHandler.cs" />
     <Compile Include="Handlers\Files\LegacyTitlePicMigration.cs" />
     <Compile Include="Handlers\Files\ThumbnailImages.cs" />
-    <Compile Include="Handlers\RegistryService.cs" />
+    <Compile Include="GameStores\Steam\SteamRegistry.cs" />
     <Compile Include="Handlers\Sync\StartupImageSyncAction.cs" />
     <Compile Include="Handlers\Sync\Doom64TitlePicSyncAction.cs" />
     <Compile Include="Handlers\Sync\GameConfSyncAction.cs" />
