+++ DoomLauncher/GameStores/Steam/SteamRegistry.cs
@@ -1,9 +1,9 @@
 ﻿using Microsoft.Win32;
 using System;
 
-namespace DoomLauncher.Handlers
+namespace DoomLauncher.GameStores.Steam
 {
-    public static class RegistryService
+    public static class SteamRegistry
     {
         // As documented at https://help.steampowered.com/en/faqs/view/3C73-90F9-F600-0266
         private const string STEAM_REGISTRY_KEY_32 = @"HKEY_LOCAL_MACHINE\SOFTWARE\Wow6432Node\Valve\Steam";
